Move to extension method
