+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Extensions/TagHelperDescriptorExtensions.cs
@@ -3,6 +3,7 @@
 
 using System;
 using Microsoft.AspNetCore.Razor.Language;
+using Microsoft.AspNetCore.Razor.PooledObjects;
 
 namespace Microsoft.CodeAnalysis.Razor.Workspaces;
 
@@ -12,5 +13,41 @@ internal static class TagHelperDescriptorExtensions
     {
         return d.Metadata.TryGetValue(TagHelperMetadata.Common.ClassifyAttributesOnly, out var value) ||
             string.Equals(value, bool.TrueString, StringComparison.OrdinalIgnoreCase);
+    }
+
+    public static string? TryGetComponentTag(this TagHelperDescriptor descriptor)
+    {
+        var typeName = descriptor.GetTypeNameIdentifier();
+        if (string.IsNullOrWhiteSpace(typeName))
+        {
+            return null;
+        }
+
+        // TODO: Add @using statements if required, or fully qualify (GetTypeName())
+
+        using var _ = StringBuilderPool.GetPooledObject(out var sb);
+
+        sb.Append('<');
+        sb.Append(typeName);
+
+        foreach (var requiredAttribute in descriptor.EditorRequiredAttributes)
+        {
+            sb.Append(' ');
+            sb.Append(requiredAttribute.Name);
+            sb.Append("=\"\"");
+        }
+
+        if (descriptor.AllowedChildTags.Length > 0)
+        {
+            sb.Append("></");
+            sb.Append(typeName);
+            sb.Append('>');
+        }
+        else
+        {
+            sb.Append(" />");
+        }
+
+        return sb.ToString();
     }
 }
